src/mcp-servers/ladbs/pyproject.toml
[project]
name = "mcp-ladbs"
version = "0.1.0"
description = "LADBS (Los Angeles Department of Building and Safety) MCP Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.19.0",
    "fastmcp>=0.3.0",
    "mcp>=1.1.2",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
